Dedicated folder for E0
This commit edits the E0.md file, specifically the line that describes
the first patch, to prompt students to put their E0.txt file inside of a
directory username/E0, instead of directly in their username directory
to be consistent with other assignments. Later in the assignment
instructions it prompts the user the put their E0.txt file inside of
their username/E0 directory, but not in the beginning of the assignment
under the "What to submit:" section.

Signed-off-by: Isaiah Andrade <isaiah_andrade@fall2024-uml.kdlp.underground.software>